Crate ruma_events_macros

A procedural macro for generating ruma-events events.

See the documentation for the individual macros for usage details.

Macros

event_enum
Generates an enum to represent the various Matrix event types.

Derive Macros

BasicEventContent
Generates an implementation of ruma_events::BasicEventContent and it’s super traits.

EphemeralRoomEventContent
Generates an implementation of ruma_events::EphemeralRoomEventContent and it’s super traits.

Event
Generates implementations needed to serialize and deserialize Matrix events.

EventContent
Generates an implementation of ruma_events::EventContent.

MessageEventContent
Generates an implementation of ruma_events::MessageEventContent and it’s super traits.

RoomEventContent
Generates an implementation of ruma_events::RoomEventContent and it’s super traits.

StateEventContent
Generates an implementation of ruma_events::StateEventContent and it’s super traits.